pyproject.toml
[tool.poetry]
name = "function-space-gradient-flow"
version = "0.0.0"
description = "Implementation of Projected Wasserstein Gradient Flows in Function Spaces"
authors = [
    "James Wu <jswu18@gmail.com>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.11"

# Core dependencies
matplotlib = "^3.6.2"
gpytorch = "^1.11"
pandas = "^2.1.1"
torch = "^2.1.0"
sklearn = "^0.0.post10"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
black = {extras = ["jupyter"], version = "^23.9.1"}
tqdm = "^4.65.0"
ipykernel = "^6.20.1"
isort = "^5.12"
mock = "^5.0.2"
pytest = "^7.4"
pytest-cov = "^4.1.0"

[[tool.poetry.source]]
name = "PyPI"
priority = "primary"

[tool.isort]
skip = [".venv"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
